+++ src/config.ts
@@ -5,22 +5,21 @@ import { createDefaultImbues } from "@data/imbues";
 import { MaterialData, materialDataSchema } from "@data/material";
 import { loadHomebrewMaterials } from "./homebrew";
 import { t } from "./utils";
-import { NormalizedValue } from "@localTypes/global";
 
 export type MonsterPartsConfig = {
     materials: Map<MaterialKey, MaterialData>;
     thresholds: {
         refinement: {
-            weapon: NormalizedValue[];
-            armor: NormalizedValue[];
-            shield: NormalizedValue[];
-            equipment: NormalizedValue[];
+            weapon: number[];
+            armor: number[];
+            shield: number[];
+            equipment: number[];
         };
         imbuement: {
-            weapon: NormalizedValue[];
-            armor: NormalizedValue[];
-            shield: NormalizedValue[];
-            equipment: NormalizedValue[];
+            weapon: number[];
+            armor: number[];
+            shield: number[];
+            equipment: number[];
         };
     };
     valueForMonsterLevel: number[];